# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 八字排盘 App API
  version: 0.1.0
  description: 八字排盘与命书生成的前后端接口契约。
...
servers:
  - url: /api/v1
paths:
...
  /reports:
    post:
      summary: 使用会员权益创建文字型命书（需 Bearer token）
      description: 仅有效会员可直接调用；会员每天最多成功生成三份。非会员返回 REPORT_PAYMENT_REQUIRED，必须改走 /reports/checkout，不能用本接口绕过购买。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportPreviewRequest'
      responses:
        '200':
          description: 报告生成记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportInfo'
        '400':
          description: 非会员需购买（REPORT_PAYMENT_REQUIRED）、会员当日额度已用完（MEMBER_DAILY_REPORT_LIMIT）或内容生成失败。
    get:
      summary: 我的命书列表（需 Bearer token；只返回 ready 报告）
      description: locked 报告在支付完成前不会出现在列表中。
      responses:
        '200':
          description: 报告列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportInfo'
  /reports/checkout:
    post:
      summary: 创建单份命书订单和锁定报告（需 Bearer token）
      description: 金额由服务端商品目录决定，客户端不提交价格。命书先完成生成并保存为 locked；支付前 locked 报告不可通过列表或详情读取。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportPreviewRequest'
      responses:
        '200':
          description: 待支付订单与其唯一关联的锁定报告
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCheckoutInfo'
        '400':
          description: 商品未开放（REPORT_PRODUCT_UNAVAILABLE）或命书生成失败；失败时不创建可收费订单。
  /reports/checkout/{orderId}/mock-pay:
    post:
      summary: 模拟支付并解锁单份命书（仅非生产环境，需 Bearer token）
      description: 仅非生产环境可用；操作幂等，重复确认不会重复创建或解锁。成功后返回 checkout 时生成的同一份命书，而不是重新生成。
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 已变为 ready 的同一份命书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportInfo'
        '400':
          description: 订单不存在、状态异常、模拟支付被禁用或命书解锁失败。
...
  /reports/{id}:
    get:
      summary: 命书详情（需 Bearer token）
      description: 仅报告所有者可以读取 ready 报告；locked 报告统一返回 404，避免未支付内容泄露。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 报告详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportInfo'
components:
  schemas:
...
    ReportTopic:
      type: string
      enum: [overall, career, wealth, relationship]
    ReportEdition:
      type: string
      enum: [plain, professional]
      description: 当前单份商品只开放事业、财富、感情通俗版；综合通俗版仅向会员开放内容验收，尚未进入单份商品目录；所有专业版均未开放。
    ReportPreviewRequest:
      type: object
      required: [request, topic, edition]
      description: overall v1.1 和 wealth v3 不需要问卷；career 必须提交 careerContext；relationship v1 必须提交 relationshipContext。综合与财富通俗版固定输出三年；感情单身版输出今年重点与明年参考，交往中和已婚版输出三年；专业版暂未开放。综合通俗版当前仅供会员验收，不支持单份购买；历史 overall v1 报告继续可读。
      properties:
        request:
          $ref: '#/components/schemas/PaipanRequest'
        topic:
          $ref: '#/components/schemas/ReportTopic'
        edition:
          $ref: '#/components/schemas/ReportEdition'
        careerContext:
          $ref: '#/components/schemas/CareerContext'
        wealthContext:
          deprecated: true
          description: 旧版财富问卷兼容字段；可省略，wealth v2 接受但完全忽略，将在后续契约中删除。
          allOf:
            - $ref: '#/components/schemas/WealthContext'
        relationshipContext:
          $ref: '#/components/schemas/RelationshipContextRequest'
    CareerContext:
      type: object
      required: [status, goal, pace]
      description: 事业现实状态只用于解释分支与行动建议，不计入命理证据。
      properties:
        status:
          type: string
          enum: [employed, self_employed, job_seeking, studying]
        goal:
          type: string
          enum: [promotion, job_change, stability, transition]
        pace:
          type: string
          enum: [smooth, stalled, high_pressure, preparing_change]
    WealthContext:
      type: object
      required: [incomeSource, goal, pace]
      deprecated: true
      description: wealth-narrative-v1 的历史字段；wealth v2 不读取、不保存为用户答案，也不允许其影响证据、排序或文案。
      properties:
        incomeSource:
          type: string
          enum: [salary, self_employed, mixed, unstable]
        goal:
          type: string
          enum: [increase_income, stabilize_cashflow, reduce_pressure, new_income_source]
        pace:
          type: string
          enum: [stable, income_fluctuating, spending_pressure, preparing_adjustment]
    RelationshipContextRequest:
      type: object
      required: [status]
      description: 感情状态只决定通俗文案的问题角度与行动表达，不参与证据提取、权重、排序、风险或年度趋势计算。
      properties:
        status:
          type: string
          enum: [single, dating, married]
          description: single=单身或尚未确定关系；dating=已确认交往关系；married=已婚或长期共同生活。
    ReportCheckoutInfo:
      type: object
      required: [orderId, reportId, topic, edition, amountCents, status]
      properties:
        orderId:
          type: integer
        reportId:
          type: integer
          description: checkout 时已经生成并锁定的报告 ID；支付成功后解锁的仍是此 ID。
        topic:
          $ref: '#/components/schemas/ReportTopic'
        edition:
          $ref: '#/components/schemas/ReportEdition'
        amountCents:
          type: integer
          readOnly: true
          description: 服务端商品目录确定的人民币分值；当前开放商品均为 690。
        status:
          type: string
          enum: [pending]
    ReportInfo:
      type: object
      required: [id, subject, topic, edition, status, contentVersion, content, createdAt, generatedAt]
      properties:
        id: { type: integer }
        subject: { type: string }
        topic: { $ref: '#/components/schemas/ReportTopic' }
        edition: { $ref: '#/components/schemas/ReportEdition' }
        status: { type: string, enum: [ready] }
        contentVersion:
          type: string
          description: career-narrative-v1/v2/v3、wealth-narrative-v1 使用旧内容结构；wealth-narrative-v2/v3 与 relationship-narrative-v1 使用各自独立结构。新生成的财富通俗版为 wealth-narrative-v3。
        content:
          oneOf:
            - $ref: '#/components/schemas/CareerNarrativePlan'
            - $ref: '#/components/schemas/WealthNarrativePlan'
            - $ref: '#/components/schemas/RelationshipNarrativePlan'
            - $ref: './drafts/wealth-v3.schema.json#/definitions/Content'
          description: 由 contentVersion 判别；旧事业及财富 v1 使用 CareerNarrativePlan，财富 v2 使用 WealthNarrativePlan，财富 v3 使用独立 Content，感情 v1 使用 RelationshipNarrativePlan。
        createdAt: { type: string, format: date-time }
        generatedAt: { type: string, format: date-time }
    CareerNarrativePlan:
      type: object
      required: [thesis, contextSummary, years, route]
      properties:
        thesis: { type: string }
        contextSummary: { type: string }
        years:
          type: array
          items: { $ref: '#/components/schemas/CareerYearNarrative' }
        route:
          type: array
          items: { type: string }
    CareerYearNarrative:
      type: object
      properties:
        year: { type: integer }
        ganZhi: { type: string }
        stage: { type: string }
        headline: { type: string }
        verdict: { type: string }
        reasons:
          type: array
          items: { type: string }
        obstacle: { type: string }
        actions:
          type: array
          items: { type: string }
        changeCondition: { type: string }
        evidenceKeys:
          type: array
          items: { type: string }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/ReportEvidence' }
        counterEvidence:
          type: array
          items: { $ref: '#/components/schemas/ReportEvidence' }
        confidence: { type: string }
    WealthPathCode:
      type: string
      enum:
        - stable_income
        - skill_income
        - project_income
        - cooperation_income
        - retention
    WealthNarrativePlan:
      type: object
      description: wealth-narrative-v2 的三年通俗版快照。产品固定三年；horizonYears 只在响应中说明快照长度，不是请求参数。
      required:
        - horizonYears
        - thesis
        - summary
        - paths
        - primaryPathCode
        - secondaryPathCode
        - mainRisk
        - years
        - route
      properties:
        horizonYears:
          type: integer
          enum: [3]
          readOnly: true
        thesis: { type: string }
        summary: { type: string }
        paths:
          type: array
          minItems: 5
          maxItems: 5
          description: 必须且仅包含五个 WealthPathCode，各出现一次。
          items: { $ref: '#/components/schemas/WealthPathSummary' }
        primaryPathCode: { $ref: '#/components/schemas/WealthPathCode' }
        secondaryPathCode: { $ref: '#/components/schemas/WealthPathCode' }
        mainRisk: { $ref: '#/components/schemas/WealthRiskSummary' }
        years:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: '#/components/schemas/WealthYearNarrative' }
        route:
          type: array
          minItems: 3
          maxItems: 3
          items: { type: string }
    WealthPathSummary:
      type: object
      required: [code, label, status, judgment, supportingEvidenceKeys, limitingEvidenceKeys]
      properties:
        code: { $ref: '#/components/schemas/WealthPathCode' }
        label: { type: string }
        status:
          type: string
          enum: [主要方向, 辅助方向, 需要留意, 表现一般]
        judgment: { type: string }
        supportingEvidenceKeys:
          type: array
          uniqueItems: true
          items: { type: string }
        limitingEvidenceKeys:
          type: array
          uniqueItems: true
          items: { type: string }
    WealthRiskSummary:
      type: object
      required: [pathCode, label, judgment, evidenceKeys]
      properties:
        pathCode: { $ref: '#/components/schemas/WealthPathCode' }
        label: { type: string }
        judgment: { type: string }
        evidenceKeys:
          type: array
          uniqueItems: true
          items: { type: string }
    WealthYearNarrative:
      type: object
      required:
        - year
        - ganZhi
        - focus
        - incomeSource
        - retention
        - mainLimit
        - realitySignals
        - actions
        - transition
        - evidenceKeys
      properties:
        year: { type: integer }
        ganZhi: { type: string }
        focus: { type: string }
        incomeSource: { type: string }
        retention: { type: string }
        mainLimit: { type: string }
        realitySignals:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: string }
        actions:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: string }
        transition: { type: string }
        evidenceKeys:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
    RelationshipStatus:
      type: string
      enum: [single, dating, married]
    RelationshipDimensionCode:
      type: string
      enum: [connection, response, daily_cooperation, boundaries, stability]
    RelationshipNarrativePlan:
      type: object
      description: relationship-narrative-v1 的三年通俗版快照。规则权重只表示规则在本报告中的方向与显著程度，不是统计概率，也不保证现实事件发生。
      required:
        - relationshipStatus
        - relationshipStatusLabel
        - horizonYears
        - thesis
        - summary
        - dimensions
        - primaryDimensionCode
        - secondaryDimensionCode
        - focusTied
        - years
        - evidenceKeys
      properties:
        relationshipStatus: { $ref: '#/components/schemas/RelationshipStatus' }
        relationshipStatusLabel: { type: string }
        horizonYears:
          type: integer
          enum: [3]
          readOnly: true
        thesis: { type: string }
        summary: { type: string }
        dimensions:
          type: array
          minItems: 5
          maxItems: 5
          items: { $ref: '#/components/schemas/RelationshipDimensionSummary' }
        primaryDimensionCode: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        secondaryDimensionCode: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        focusTied: { type: boolean }
        mainRisk:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RelationshipRiskSummary'
        years:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: '#/components/schemas/RelationshipYearNarrative' }
        evidenceKeys:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
    RelationshipDimensionSummary:
      type: object
      required: [code, label, status, tone, judgment, supportingEvidenceKeys, limitingEvidenceKeys]
      properties:
        code: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        label: { type: string }
        status: { type: string }
        tone:
          type: string
          enum: [supportive, mixed, pressured, quiet]
        judgment: { type: string }
        supportingEvidenceKeys:
          type: array
          uniqueItems: true
          items: { type: string }
        limitingEvidenceKeys:
          type: array
          uniqueItems: true
          items: { type: string }
    RelationshipRiskSummary:
      type: object
      required: [dimensionCode, label, judgment, evidenceKeys]
      properties:
        dimensionCode: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        label: { type: string }
        judgment: { type: string }
        evidenceKeys:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
    RelationshipYearNarrative:
      type: object
      required:
        - year
        - ganZhi
        - focus
        - judgment
        - realitySignals
        - actions
        - transition
        - primaryDimensionCode
        - secondaryDimensionCode
        - evidenceKeys
      properties:
        year: { type: integer }
        ganZhi: { type: string }
        focus: { type: string }
        judgment: { type: string }
        mainLimit:
          type: string
          nullable: true
          description: 仅在该年存在真实负面证据时出现。
        realitySignals:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: string }
        actions:
          type: array
          minItems: 2
          maxItems: 2
          uniqueItems: true
          items: { type: string }
        transition: { type: string }
        primaryDimensionCode: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        secondaryDimensionCode: { $ref: '#/components/schemas/RelationshipDimensionCode' }
        riskDimensionCode:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RelationshipDimensionCode'
        evidenceKeys:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
    ReportEvidence:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        value: { type: string }
    PaipanRequest:
      type: object
      required: [gender, solarDateTime, trueSolarTime]
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [male, female]
        solarDateTime:
          type: string
          description: ISO 格式公历时间，如 1995-10-08T14:30:00
        birthPlace:
          type: string
          description: 出生地，用于真太阳时
        trueSolarTime:
          type: boolean
...